refactor: extract duplicate escape_toml_string to commands/toml_utils
The same pure TOML-string escape function was defined verbatim in both
commands/publish.rs and commands/deps/lockfile.rs; the duplication was
acknowledged in a comment at publish.rs:562 with a weak rationale.

Consolidate into a new pub(crate) commands/toml_utils module. Both call
sites now import from there; the two-test suite moves to toml_utils as
well. No behaviour change.
